# Tattel.WebApiClient/Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Tattel Web Api",
    "description": "Tattel Web Service",
    "contact": {
      "name": "Tattel",
      "email": "tattelglobal@gmail.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/UserLocation/{id}": {
      "get": {
        "tags": [
          "UserLocation"
        ],
        "operationId": "GetUserLocationLocation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLocation"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "UserLocation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          },
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/UserLocation": {
      "get": {
        "tags": [
          "UserLocation"
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserLocation"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "UserLocation"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UserLocation"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLocation"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLocation"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserLocation"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          },
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "UserLocation"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UserLocation"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLocation"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLocation"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserLocation"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          },
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "UserLocation"
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          },
          "204": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GPSLocation": {
        "type": "object",
        "properties": {
          "locationId": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "UserLocation": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "gpsLocation": {
            "$ref": "#/components/schemas/GPSLocation"
          }
        },
        "additionalProperties": false
      }
    }
  }
}